Support for commands to change camera FPSs, plus a quick commander utility
