+++ .gitignore
@@ -4,4 +4,4 @@
 dv??
 grasp
 main
-fake_tm
+quick
